remove OdbNodeTp3 -> we only need NodeRefTp3
